+++ src/events/ModifyFormIntegrationsEvent.php
@@ -1,0 +1,14 @@
+<?php
+namespace verbb\formie\events;
+
+use yii\base\Event;
+
+class ModifyFormIntegrationsEvent extends Event
+{
+    // Properties
+    // =========================================================================
+
+    public $integrations;
+    public $type;
+    
+}